requirements.txt
Flask
langchain-ollama
langchain-community
faiss-cpu
python-dotenv
pdfplumber
